mostly running init proc (go)

+++ include/kernel/queue.h
@@ -26,7 +26,8 @@ struct mq_open_params
 	char name[MAX_MQ_NAME_SIZE];
 };
 
-struct mq_send_params {
+struct mq_send_params
+{
 	uint32_t flags;
 	uint32_t id;
 	char name[MAX_MQ_NAME_SIZE];
@@ -82,7 +83,8 @@ typedef struct queue_list_entry_t
 	struct list_head queues;
 } queue_list_entry_t;
 
-typedef struct queue_buffer_t {
+typedef struct queue_buffer_t
+{
 	spinlock_t lock;
 	uint32_t refs;
 
@@ -93,7 +95,8 @@ typedef struct queue_buffer_t {
 	const char buf[];
 } queue_buffer_t;
 
-typedef struct queue_buffer_list_entry_t {
+typedef struct queue_buffer_list_entry_t
+{
 	struct list_head list;
 
 	queue_buffer_t *buffer;
@@ -103,16 +106,16 @@ void queues_init();
 
 skiplist_t *queues_get_skl();
 
-int syscall_mq_open(thread_t *thread, ...);
+uint64_t syscall_mq_open(thread_t *thread, ...);
 
-int syscall_mq_close(thread_t *thread, ...);
+uint64_t syscall_mq_close(thread_t *thread, ...);
 
-int syscall_mq_ctrl(thread_t *thread, ...);
+uint64_t syscall_mq_ctrl(thread_t *thread, ...);
 
-int syscall_mq_send(thread_t *thread, ...);
+uint64_t syscall_mq_send(thread_t *thread, ...);
 
-int syscall_mq_recv(thread_t *thread, ...);
+uint64_t syscall_mq_recv(thread_t *thread, ...);
 
-queue_list_entry_t* queues_find_by_entry(queue_list_entry_t* sq);
+queue_list_entry_t *queues_find_by_entry(queue_list_entry_t *sq);
 
 #endif
